+++ test/types/populate.test.ts
@@ -7,7 +7,7 @@ interface Child {
   name: string;
 }
 
-const childSchema: Schema = new Schema({ name: String });
+const childSchema = new Schema({ name: String });
 const ChildModel = model<Child>('Child', childSchema);
 
 interface Parent {
@@ -119,7 +119,7 @@ function gh11014() {
       name: String
     })
   );
-  const childSchema: Schema = new Schema({ name: String });
+  const childSchema = new Schema({ name: String });
   const ChildModel = model<Child>('Child', childSchema);
 
   // Populate with `Paths` generic `{ child: Child }` to override `child` path
@@ -267,7 +267,7 @@ async function gh11710() {
     child: { type: Schema.Types.ObjectId, ref: 'Child' },
     name: String
   }));
-  const childSchema: Schema = new Schema({ name: String });
+  const childSchema = new Schema({ name: String });
   const ChildModel = model<Child>('Child', childSchema);
 
   // Populate with `Paths` generic `{ child: Child }` to override `child` path
@@ -280,7 +280,7 @@ async function gh11758() {
     name: string
     _id: Types.ObjectId
   }
-  const nestedChildSchema: Schema = new Schema({ name: String });
+  const nestedChildSchema = new Schema({ name: String });
 
   interface Parent {
     nestedChild: Types.ObjectId
@@ -319,7 +319,7 @@ async function gh11955() {
   interface Child {
     name: string;
   }
-  const childSchema: Schema = new Schema({ name: String });
+  const childSchema = new Schema({ name: String });
   model<Child>('Child', childSchema);
 
   const parent = await ParentModel.findOne({}).exec();
@@ -389,7 +389,7 @@ function gh14441() {
   interface Child {
     name: string;
   }
-  const childSchema: Schema = new Schema({ name: String });
+  const childSchema = new Schema({ name: String });
   model<Child>('Child', childSchema);
 
   ParentModel.findOne({})